fixes large number of reads issue (part 2)
For now we leave the dependency array of the useState in Homepage.js empty. For some reason when we add postsCollectionRef to the dependency array it causes the issue (I'm not sure why). So this will work for now. Still to a lot to do with the Firestore implementation though